✨(frontend) create JoanieSessionProvider
When Joanie is enabled, we would like to retrieve in background information
related to the authenticated user so we need to make this request when
SessionProvider is mount and user session is retrieved.
Furthermore to avoid to load unused code when Joanie is disabled, we lazy load
the right SessionProvider according to the context.
Finaly to clarify structure, folder containing this code
has been renamed SessionProvider.